Extract assertions from test case.
