+++ chevah/compat/testing/testcase.py
@@ -7,12 +7,11 @@ TestCase used for Chevah project.
 from __future__ import print_function
 from __future__ import division
 from __future__ import absolute_import
-from builtins import next
 from builtins import str
 from builtins import range
 from builtins import object
 from contextlib import contextmanager
-import collections
+
 import inspect
 import threading
 import os
@@ -42,8 +41,8 @@ from chevah.compat import (
     system_users,
     SuperAvatar,
     )
-from chevah.compat.exceptions import CompatError
 from chevah.compat.administration import os_administration
+from chevah.compat.testing.assertion import AssertionMixin
 from chevah.compat.testing.mockup import mk
 from chevah.compat.testing.constants import (
     TEST_NAME_MARKER,
@@ -60,12 +59,6 @@ else:
     from unittest import TestCase
 
 try:
-    from zope.interface.verify import verifyObject
-except ImportError:
-    # Zope support is optional.
-    pass
-
-try:
     # Import reactor last in case some other modules are changing the reactor.
     from twisted.internet import reactor
 except ImportError:
@@ -77,14 +70,6 @@ def _get_hostname():
     Return hostname as resolved by default DNS resolver.
     """
     return socket.gethostname()
-
-
-class Contains(object):
-    """
-    Marker class used in tests when something needs to contain a value.
-    """
-    def __init__(self, value):
-        self.value = value
 
 
 class TwistedTestCase(TestCase):
@@ -733,7 +718,7 @@ class TwistedTestCase(TestCase):
                 'Deferred contains a failure: %s' % (error))
 
 
-class ChevahTestCase(TwistedTestCase):
+class ChevahTestCase(TwistedTestCase, AssertionMixin):
     """
     Test case for Chevah tests.
 
@@ -748,7 +733,6 @@ class ChevahTestCase(TwistedTestCase):
     hostname = _get_hostname()
 
     Bunch = Bunch
-    Contains = Contains
     Mock = Mock
     #: Obsolete. Please use self.patch and self.patchObject.
     Patch = patch
@@ -1033,7 +1017,7 @@ class ChevahTestCase(TwistedTestCase):
 
     def assertRaises(self, exception_class, callback=None, *args, **kwargs):
         """
-        Wrapper around the stdlib call to allow non-context usage.
+        Overwrites the stdlib call to allow non-context usage.
         """
         super_assertRaises = super(ChevahTestCase, self).assertRaises
         if callback is None:
@@ -1043,322 +1027,6 @@ class ChevahTestCase(TwistedTestCase):
             callback(*args, **kwargs)
 
         return context.exception
-
-    def assertCompatError(self, expected_id, actual_error):
-        """
-        Raise an error if `actual_error` is not a `CompatError` instance.
-
-        Raise an error if `expected_id` does not match event_id of
-        `actual_error`.
-        """
-        if not isinstance(actual_error, CompatError):
-            values = (actual_error, type(actual_error))
-            message = u'Error %s not CompatError but %s' % values
-            raise AssertionError(message.encode('utf-8'))
-
-        actual_id = getattr(actual_error, 'event_id', None)
-        if expected_id != actual_id:
-            values = (actual_error, str(expected_id), str(actual_id))
-            message = u'Error id for %s is not %s, but %s.' % values
-            raise AssertionError(message.encode('utf-8'))
-
-    @classmethod
-    def assertTempIsClean(cls):
-        """
-        Raise an error if the temporary folder contains any testing
-        specific files for folders.
-        """
-        members = cls.cleanTemporaryFolder()
-        if members:
-            message = u'Temporary folder is not clean. %s' % (
-                u', '.join(members))
-            raise AssertionError(message.encode('utf-8'))
-
-    @classmethod
-    def assertWorkingFolderIsClean(cls):
-        """
-        Raise an error if the current working folder contains any testing
-        specific files for folders.
-        """
-        members = cls.cleanWorkingFolder()
-        if members:
-            message = u'Working folder is not clean. %s' % (
-                u', '.join(members))
-            raise AssertionError(message.encode('utf-8'))
-
-    def assertIsFalse(self, value):
-        '''Raise an exception if value is not 'False'.'''
-        if value is not False:
-            raise AssertionError('%s is not False.' % str(value))
-
-    def assertIsTrue(self, value):
-        '''Raise an exception if value is not 'True'.'''
-        if value is not True:
-            raise AssertionError('%s is not True.' % str(value))
-
-    def assertIsInstance(self, expected_type, value, msg=None):
-        """
-        Raise an exception if `value` is not an instance of `expected_type`
-        """
-        # In Python 2.7 isInstance is already defined, but with swapped
-        # arguments.
-        if not inspect.isclass(expected_type):
-            expected_type, value = value, expected_type
-
-        if not isinstance(value, expected_type):
-            raise AssertionError(
-                "Expecting type %s, but got %s. %s" % (
-                    expected_type, type(value), msg))
-
-    def assertIsListening(self, ip, port, debug=False, clear_log=False):
-        '''Check if the port and address are in listening mode.'''
-        test_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        test_socket.settimeout(1.0)
-        try:
-            test_socket.connect((ip, port))
-            sock_name = test_socket.getsockname()
-            test_socket.shutdown(socket.SHUT_RDWR)
-            if debug:
-                print('Connected as: %s:%d' % (sock_name[0], sock_name[1]))
-        except:
-            raise AssertionError(
-                'It seems that no one is listening on %s:%d' % (
-                    ip, port))
-        if clear_log:
-            # Clear the log since we don't care about log generated by
-            # assertIsListening.
-            # We need to wait a bit.
-            time.sleep(0.1)
-            self.clearLog()
-
-    def assertIsNotListening(self, ip, port):
-        '''Check if the port and address are in listening mode.'''
-        try:
-            self.assertIsListening(ip, port)
-        except AssertionError:
-            return
-        raise AssertionError(
-            'It seems that someone is listening on %s:%d' % (
-                ip, port))
-
-    def assertEqual(self, first, second, msg=None):
-        '''Extra checks for assert equal.'''
-        try:
-            super(ChevahTestCase, self).assertEqual(first, second, msg)
-        except AssertionError as error:
-            message = error.message
-            if isinstance(message, str):
-                message = message.encode('utf-8')
-            raise AssertionError(message)
-
-        if (isinstance(first, str) and not isinstance(second, str)):
-            if not msg:
-                msg = u'Type of "%s" is unicode while for "%s" is str.' % (
-                    first, second)
-            raise AssertionError(msg.encode('utf-8'))
-
-        if (not isinstance(first, str) and isinstance(second, str)):
-            if not msg:
-                msg = u'Type of "%s" is str while for "%s" is unicode.' % (
-                    first, second)
-            raise AssertionError(msg.encode('utf-8'))
-
-    def assertFailureType(self, failure_class, failure_or_deferred):
-        '''Raise assertion error if failure is not of required type.'''
-        if isinstance(failure_or_deferred, Failure):
-            failure = failure_or_deferred
-        else:
-            self.assertIsFailure(failure_or_deferred)
-            failure = failure_or_deferred.result
-
-        if failure.type is not failure_class:
-            message = u'Failure %s is not of type %s' % (
-                str(failure), failure_class)
-            raise AssertionError(message.encode('utf-8'))
-
-    def assertFailureID(self, failure_id, failure_or_deferred):
-        """
-        Raise `AssertionError` if failure does not have the required id or
-        the specified id is not unicode.
-        """
-        if isinstance(failure_or_deferred, Failure):
-            failure = failure_or_deferred
-        else:
-            self.assertIsFailure(failure_or_deferred)
-            failure = failure_or_deferred.result
-
-        try:
-            actual_id = getattr(failure.value, 'id')
-        except:
-            actual_id = getattr(failure.value, 'event_id')
-
-        if not isinstance(actual_id, str):
-            raise AssertionError('Failure ID must be unicode.')
-
-        if actual_id != failure_id:
-            message = u'Failure id for %s is not %s, but %s' % (
-                failure, str(failure_id), str(actual_id))
-            raise AssertionError(message.encode('utf-8'))
-
-    def assertFailureData(self, data, failure_or_deferred):
-        """
-        Raise AssertionError if failure does not contain the required data.
-        """
-        if isinstance(failure_or_deferred, Failure):
-            failure = failure_or_deferred
-        else:
-            self.assertIsFailure(failure_or_deferred)
-            failure = failure_or_deferred.result
-
-        failure_data = failure.value.data
-        try:
-            failure_id = getattr(failure.value, 'id')
-        except:
-            failure_id = getattr(failure.value, 'event_id')
-
-        self._checkData(
-            kind=u'Failure',
-            kind_id=failure_id,
-            expected_data=data,
-            current_data=failure_data,
-            )
-
-    def _checkData(self, kind, kind_id, expected_data, current_data):
-        """
-        Helper for sharing same code between various data checkers.
-        """
-        for key, value in expected_data.items():
-            try:
-                current_value = current_data[key]
-
-                if isinstance(value, Contains):
-                    if value.value not in current_value:
-                        message = (
-                            u'%s %s, for data "%s" does not contains "%s", '
-                            u'but is "%s"') % (
-                            kind, str(kind_id), key, value.value,
-                            current_value)
-                        raise AssertionError(message.encode('utf-8'))
-                else:
-                    if value != current_value:
-                        message = (
-                            u'%s %s, for data "%s" is not "%s", but "%s"') % (
-                            kind,
-                            str(kind_id),
-                            key,
-                            repr(value),
-                            repr(current_value),
-                            )
-                        raise AssertionError(message.encode('utf-8'))
-            except KeyError:
-                values = (
-                    kind, str(kind_id), repr(key), repr(current_data))
-                message = u'%s %s, has no data "%s". Data is:\n%s' % values
-                raise AssertionError(message.encode('utf-8'))
-
-    def assertIsEmpty(self, target):
-        """
-        Raise AssertionError if target is not empty.
-        """
-        if isinstance(target, collections.Iterable):
-            iterator = iter(target)
-            try:
-                next(iterator)
-            except StopIteration:
-                pass
-            else:
-                message = u'Iterable is not empty.\n%s.' % (target,)
-                raise AssertionError(message.encode('utf-8'))
-            return
-
-        if len(target) != 0:
-            message = u'Value is not empty.\n%s.' % (target)
-            raise AssertionError(message.encode('utf-8'))
-
-    def assertIsNotEmpty(self, target):
-        """
-        Raise AssertionError if target is empty.
-        """
-        if isinstance(target, collections.Iterable):
-            try:
-                self.assertIsEmpty(target)
-            except AssertionError:
-                pass
-            else:
-                message = u'Iterable is empty.\n%s.' % target
-                raise AssertionError(message.encode('utf-8'))
-            return
-
-        if len(target) == 0:
-            raise AssertionError('Value is empty.\n%s.' % (target))
-
-    def assertIn(self, target, source):
-        """
-        Raise AssertionError if source is not in target.
-        """
-        if source not in target:
-            message = u'%s not in %s.' % (repr(source), repr(target))
-            raise AssertionError(message.encode('utf-8'))
-
-    def assertContains(self, token, source):
-        """
-        Raise AssertionError if source does not contain `token`.
-        """
-        if token not in source:
-            message = u'%s does not contains %s.' % (
-                repr(source), repr(token))
-            raise AssertionError(message.encode('utf-8'))
-
-    def assertNotContains(self, token, source):
-        """
-        Raise AssertionError if source does contain `token`.
-        """
-        if token in source:
-            message = u'%s contains %s.' % (repr(source), repr(token))
-            raise AssertionError(message.encode('utf-8'))
-
-    def assertTextContains(self, pattern, source):
-        """
-        Raise AssertionError if pattern is not found in source.
-        """
-        if pattern not in pattern:
-            message = u'%s not contained in\n%s.' % (
-                repr(pattern), repr(source))
-            raise AssertionError(message.encode('utf-8'))
-
-    def assertStartsWith(self, start, source):
-        """
-        Raise AssertionError if `source` does not starts with `start`.
-        """
-        if not source.startswith(start):
-            message = u'%s does not starts with %s' % (
-                repr(source), repr(start))
-            raise AssertionError(message.encode('utf-8'))
-
-    def assertEndsWith(self, end, source):
-        """
-        Raise AssertionError if `source` does not ends with `end`.
-        """
-        if not source.endswith(end):
-            message = u'%s does not end with %s' % (repr(source), repr(end))
-            raise AssertionError(message.encode('utf-8'))
-
-    def assertProvides(self, interface, obj):
-        self.assertTrue(
-            interface.providedBy(obj),
-            'Object %s does not provided interface %s.' % (obj, interface))
-        verifyObject(interface, obj)
-
-    def assertNotProvides(self, interface, obj):
-        self.assertFalse(
-            interface.providedBy(obj),
-            'Object %s does not provided interface %s.' % (obj, interface))
-
-    def assertImplements(self, interface, klass):
-        self.assertTrue(
-            interface.implementedBy(klass),
-            u'Class %s does not implements interface %s.' % (
-                klass, interface))
 
 
 class FileSystemTestCase(ChevahTestCase):
